Remove redundant casts of literal offsets (`\.offset\(([0-9]+) as libc::c_int as isize\)` => `.offset($1)`).
These are literals and offsets must be `isize`,
so there can't be any type ambiguity with the casts removed.

+++ src/cdef_apply_tmpl_16.rs
@@ -754,12 +754,12 @@ unsafe extern "C" fn backup2lines(
     mut stride: *const ptrdiff_t,
     layout: Dav1dPixelLayout,
 ) {
-    let y_stride: ptrdiff_t = PXSTRIDE(*stride.offset(0 as libc::c_int as isize));
+    let y_stride: ptrdiff_t = PXSTRIDE(*stride.offset(0));
     if y_stride < 0 {
         memcpy(
-            (*dst.offset(0 as libc::c_int as isize)).offset(y_stride as isize)
+            (*dst.offset(0)).offset(y_stride as isize)
                 as *mut libc::c_void,
-            (*src.offset(0 as libc::c_int as isize))
+            (*src.offset(0))
                 .offset((7 as libc::c_int as isize * y_stride) as isize)
                 as *const libc::c_void,
             (-(2 as libc::c_int) as isize * y_stride << 1 as libc::c_int)
@@ -767,8 +767,8 @@ unsafe extern "C" fn backup2lines(
         );
     } else {
         memcpy(
-            *dst.offset(0 as libc::c_int as isize) as *mut libc::c_void,
-            (*src.offset(0 as libc::c_int as isize))
+            *dst.offset(0) as *mut libc::c_void,
+            (*src.offset(0))
                 .offset(6  * y_stride as isize)
                 as *const libc::c_void,
             (2 * y_stride << 1 as libc::c_int)
@@ -776,7 +776,7 @@ unsafe extern "C" fn backup2lines(
         );
     }
     if layout as libc::c_uint != DAV1D_PIXEL_LAYOUT_I400 as libc::c_int as libc::c_uint {
-        let uv_stride: ptrdiff_t = PXSTRIDE(*stride.offset(1 as libc::c_int as isize));
+        let uv_stride: ptrdiff_t = PXSTRIDE(*stride.offset(1));
         if uv_stride < 0 {
             let uv_off: libc::c_int = if layout as libc::c_uint
                 == DAV1D_PIXEL_LAYOUT_I420 as libc::c_int as libc::c_uint
@@ -786,18 +786,18 @@ unsafe extern "C" fn backup2lines(
                 7 as libc::c_int
             };
             memcpy(
-                (*dst.offset(1 as libc::c_int as isize)).offset(uv_stride as isize)
+                (*dst.offset(1)).offset(uv_stride as isize)
                     as *mut libc::c_void,
-                (*src.offset(1 as libc::c_int as isize))
+                (*src.offset(1))
                     .offset((uv_off as isize * uv_stride) as isize)
                     as *const libc::c_void,
                 (-(2 as libc::c_int) as isize * uv_stride << 1 as libc::c_int)
                     as size_t,
             );
             memcpy(
-                (*dst.offset(2 as libc::c_int as isize)).offset(uv_stride as isize)
+                (*dst.offset(2)).offset(uv_stride as isize)
                     as *mut libc::c_void,
-                (*src.offset(2 as libc::c_int as isize))
+                (*src.offset(2))
                     .offset((uv_off as isize * uv_stride) as isize)
                     as *const libc::c_void,
                 (-(2 as libc::c_int) as isize * uv_stride << 1 as libc::c_int)
@@ -812,16 +812,16 @@ unsafe extern "C" fn backup2lines(
                 6 as libc::c_int
             };
             memcpy(
-                *dst.offset(1 as libc::c_int as isize) as *mut libc::c_void,
-                (*src.offset(1 as libc::c_int as isize))
+                *dst.offset(1) as *mut libc::c_void,
+                (*src.offset(1))
                     .offset((uv_off_0 as isize * uv_stride) as isize)
                     as *const libc::c_void,
                 (2 * uv_stride << 1 as libc::c_int)
                     as size_t,
             );
             memcpy(
-                *dst.offset(2 as libc::c_int as isize) as *mut libc::c_void,
-                (*src.offset(2 as libc::c_int as isize))
+                *dst.offset(2) as *mut libc::c_void,
+                (*src.offset(2))
                     .offset((uv_off_0 as isize * uv_stride) as isize)
                     as *const libc::c_void,
                 (2 * uv_stride << 1 as libc::c_int) as size_t,
@@ -842,9 +842,9 @@ unsafe extern "C" fn backup2x8(
         let mut y: libc::c_int = 0 as libc::c_int;
         while y < 8 as libc::c_int {
             memcpy(
-                ((*dst.offset(0 as libc::c_int as isize))[y as usize]).as_mut_ptr()
+                ((*dst.offset(0))[y as usize]).as_mut_ptr()
                     as *mut libc::c_void,
-                &mut *(*src.offset(0 as libc::c_int as isize))
+                &mut *(*src.offset(0))
                     .offset(
                         (y_off + x_off as isize
                             - 2 as libc::c_int as isize) as isize,
@@ -852,7 +852,7 @@ unsafe extern "C" fn backup2x8(
                 ((2 as libc::c_int) << 1 as libc::c_int) as size_t,
             );
             y += 1;
-            y_off += PXSTRIDE(*src_stride.offset(0 as libc::c_int as isize));
+            y_off += PXSTRIDE(*src_stride.offset(0));
         }
     }
     if layout as libc::c_uint == DAV1D_PIXEL_LAYOUT_I400 as libc::c_int as libc::c_uint
@@ -869,7 +869,7 @@ unsafe extern "C" fn backup2x8(
     let mut y_0: libc::c_int = 0 as libc::c_int;
     while y_0 < 8 as libc::c_int >> ss_ver {
         memcpy(
-            ((*dst.offset(1 as libc::c_int as isize))[y_0 as usize]).as_mut_ptr()
+            ((*dst.offset(1))[y_0 as usize]).as_mut_ptr()
                 as *mut libc::c_void,
             &mut *(*src.offset(1))
                 .offset(
@@ -879,9 +879,9 @@ unsafe extern "C" fn backup2x8(
             ((2 as libc::c_int) << 1 as libc::c_int) as size_t,
         );
         memcpy(
-            ((*dst.offset(2 as libc::c_int as isize))[y_0 as usize]).as_mut_ptr()
+            ((*dst.offset(2))[y_0 as usize]).as_mut_ptr()
                 as *mut libc::c_void,
-            &mut *(*src.offset(2 as libc::c_int as isize))
+            &mut *(*src.offset(2))
                 .offset(
                     (y_off + x_off as isize - 2)
                         as isize,
@@ -889,7 +889,7 @@ unsafe extern "C" fn backup2x8(
             ((2 as libc::c_int) << 1 as libc::c_int) as size_t,
         );
         y_0 += 1;
-        y_off += PXSTRIDE(*src_stride.offset(1 as libc::c_int as isize));
+        y_off += PXSTRIDE(*src_stride.offset(1));
     }
 }
 unsafe extern "C" fn adjust_strength(
@@ -930,9 +930,9 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
             0 as libc::c_int
         })) as CdefEdgeFlags;
     let mut ptrs: [*mut pixel; 3] = [
-        *p.offset(0 as libc::c_int as isize),
-        *p.offset(1 as libc::c_int as isize),
-        *p.offset(2 as libc::c_int as isize),
+        *p.offset(0),
+        *p.offset(1),
+        *p.offset(2),
     ];
     let sbsz: libc::c_int = 16 as libc::c_int;
     let sb64w: libc::c_int = (*f).sb128w << 1 as libc::c_int;
@@ -1071,10 +1071,10 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                     .as_ptr()
                     .offset(by_idx as isize) as *const [uint16_t; 2];
                 noskip_mask = ((*noskip_row
-                    .offset(0 as libc::c_int as isize))[1]
+                    .offset(0))[1]
                     as libc::c_uint) << 16 as libc::c_int
                     | (*noskip_row
-                        .offset(0 as libc::c_int as isize))[0]
+                        .offset(0))[0]
                         as libc::c_uint;
                 y_lvl = (*(*f).frame_hdr).cdef.y_strength[cdef_idx as usize];
                 uv_lvl = (*(*f).frame_hdr).cdef.uv_strength[cdef_idx as usize];
@@ -1175,7 +1175,7 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                                     + (bx * 4 as libc::c_int) as isize;
                                 top = &mut *(*((*f).lf.cdef_lpf_line)
                                     .as_mut_ptr()
-                                    .offset(0 as libc::c_int as isize))
+                                    .offset(0))
                                     .offset(offset as isize) as *mut pixel;
                             } else {
                                 offset = (sby * ((4 as libc::c_int) << sb128)
@@ -1183,7 +1183,7 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                                     + (bx * 4 as libc::c_int) as isize;
                                 top = &mut *(*((*f).lf.lr_lpf_line)
                                     .as_mut_ptr()
-                                    .offset(0 as libc::c_int as isize))
+                                    .offset(0))
                                     .offset(offset as isize) as *mut pixel;
                             }
                             bot = (bptrs[0])
@@ -1196,7 +1196,7 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                                 .as_mut_ptr()
                                 .offset(tf as isize))
                                 .as_mut_ptr()
-                                .offset(0 as libc::c_int as isize))
+                                .offset(0))
                                 .offset(
                                     ((sby * 4 as libc::c_int) as isize * y_stride
                                         + (bx * 4 as libc::c_int) as isize) as isize,
@@ -1207,7 +1207,7 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                                     + (bx * 4 as libc::c_int) as isize;
                                 bot = &mut *(*((*f).lf.cdef_lpf_line)
                                     .as_mut_ptr()
-                                    .offset(0 as libc::c_int as isize))
+                                    .offset(0))
                                     .offset(offset as isize) as *mut pixel;
                             } else {
                                 let line: libc::c_int = sby * ((4 as libc::c_int) << sb128)
@@ -1216,7 +1216,7 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                                     + (bx * 4 as libc::c_int) as isize;
                                 bot = &mut *(*((*f).lf.lr_lpf_line)
                                     .as_mut_ptr()
-                                    .offset(0 as libc::c_int as isize))
+                                    .offset(0))
                                     .offset(offset as isize) as *mut pixel;
                             }
                             current_block_84 = 17075014677070940716;
@@ -1231,7 +1231,7 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                                     .as_mut_ptr()
                                     .offset(tf as isize))
                                     .as_mut_ptr()
-                                    .offset(0 as libc::c_int as isize))
+                                    .offset(0))
                                     .offset(
                                         (have_tt as isize * offset
                                             + (bx * 4 as libc::c_int) as isize) as isize,
@@ -1408,7 +1408,7 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                     }
                     bptrs[0 as libc::c_int
                         as usize] = (bptrs[0])
-                        .offset(8 as libc::c_int as isize);
+                        .offset(8);
                     bptrs[1 as libc::c_int
                         as usize] = (bptrs[1])
                         .offset((8 as libc::c_int >> ss_hor) as isize);
